@@ -12,39 +12,37 @@
         <!-- Plugin Properties -->
         <maven.compiler.parameters>true</maven.compiler.parameters>
 
-        <maven.compiler.source>17</maven.compiler.source>
-        <maven.compiler.target>17</maven.compiler.target>
+        <maven.compiler.source>21</maven.compiler.source>
+        <maven.compiler.target>21</maven.compiler.target>
 
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
 
         <!-- Wildfly Runtime Dependencies -->
-        <jakarta.jakartaee-api.version>8.0.0</jakarta.jakartaee-api.version>
-        <microprofile.version>4.1</microprofile.version>
-        <slf4j-api.version>1.7.36</slf4j-api.version>
-        <opentracing-jaxrs2.version>1.0.0</opentracing-jaxrs2.version>
-        <org.apache.cxf.version>3.4.10</org.apache.cxf.version>  <!-- unused import in api.DefaultApi, we need this at compile time, should be removed in the future-->
+        <jakarta.jakartaee-api.version>10.0.0</jakarta.jakartaee-api.version>
+        <microprofile.version>7.0</microprofile.version>
+        <slf4j-api.version>2.0.16</slf4j-api.version>
 
         <!-- Application Dependencies -->
-        <commons-lang3.version>3.12.0</commons-lang3.version>
-        <krazo-resteasy.version>1.1.0</krazo-resteasy.version>
-        <omnifaces.version>3.11</omnifaces.version>
-        <font-awesome.version>5.15.2</font-awesome.version>
-        <mdb-ui-kit.version>3.9.0</mdb-ui-kit.version>
+        <jakarta.mvc-api.version>2.1.0</jakarta.mvc-api.version>
+        <krazo-resteasy.version>3.0.1</krazo-resteasy.version>
+        <commons-lang3.version>3.17.0</commons-lang3.version>
 
-        <!-- Runtime Dependencies -->
-        <postgresql.version>42.2.27</postgresql.version>
+        <!-- Wildfly Version -->
+        <wildfly-galleon-pack.version>36.0.1.Final</wildfly-galleon-pack.version>
+        <wildfly-datasources-galleon-pack.version>10.0.0.Final</wildfly-datasources-galleon-pack.version>
 
         <!-- Plugin Versions -->
-        <dependency-plugin.version>3.3.0</dependency-plugin.version>
-        <maven-enforcer-plugin.version>3.4.1</maven-enforcer-plugin.version>
-        <openapi-generator-maven-plugin.version>6.6.0</openapi-generator-maven-plugin.version>
+        <dependency-plugin.version>3.8.1</dependency-plugin.version>
+        <maven-enforcer-plugin.version>3.5.0</maven-enforcer-plugin.version>
+        <wildfly.maven.plugin.version>5.1.3.Final</wildfly.maven.plugin.version>
+        <openapi-generator-maven-plugin.version>7.13.0</openapi-generator-maven-plugin.version>
 
         <revision>1.0.0-SNAPSHOT</revision>
     </properties>
 
     <dependencies>
-        <!-- Wildfly Runtime Dependencies -->
+        <!-- Jakarta and MicroProfile Dependencies -->
         <dependency>
             <groupId>jakarta.platform</groupId>
             <artifactId>jakarta.jakartaee-api</artifactId>
