@@ -1,57 +1,51 @@
 package com.gepardec.training.microprofile;
 
-import javax.enterprise.context.ApplicationScoped;
-import javax.inject.Named;
+import jakarta.enterprise.context.ApplicationScoped;
+import jakarta.inject.Named;
 
 @Named("mpUri")
 @ApplicationScoped
 public class MicroprofileSpecificationUri {
 
-    private static final String FAULT_TOLERANCE_URI = "https://download.eclipse.org/microprofile/microprofile-fault-tolerance-3.0/microprofile-fault-tolerance-spec-3.0.html";
+    private static final String CONFIG_URI = "https://download.eclipse.org/microprofile/microprofile-config-3.1/microprofile-config-spec-3.1.html";
 
-    private static final String JWT_URI = "https://download.eclipse.org/microprofile/microprofile-jwt-auth-1.2/microprofile-jwt-auth-spec-1.2.html";
+    private static final String FAULT_TOLERANCE_URI = "https://download.eclipse.org/microprofile/microprofile-fault-tolerance-4.1/microprofile-fault-tolerance-spec-4.1.html";
 
-    private static final String CONFIG_URI = "https://download.eclipse.org/microprofile/microprofile-config-2.0/microprofile-config-spec-2.0.html";
+    private static final String HEALTH_URI = "https://download.eclipse.org/microprofile/microprofile-health-4.0.1/microprofile-health-spec-4.0.1.html";
 
-    private static final String OPEN_API_URI = "https://download.eclipse.org/microprofile/microprofile-open-api-2.0/microprofile-openapi-spec-2.0.html";
+    private static final String TELEMETRY_URI = "https://download.eclipse.org/microprofile/microprofile-telemetry-2.0/microprofile-telemetry-spec-2.0.html";
 
-    private static final String REST_CLIENT_URI = "https://download.eclipse.org/microprofile/microprofile-rest-client-2.0/microprofile-rest-client-spec-2.0.html";
+    private static final String OPEN_API_URI = "https://download.eclipse.org/microprofile/microprofile-open-api-4.0/microprofile-openapi-spec-4.0.html";
 
-    private static final String HEALTH_URI = "https://download.eclipse.org/microprofile/microprofile-health-3.1/microprofile-health-spec-3.1.html";
+    private static final String JWT_AUTHENTICATION_URI = "https://download.eclipse.org/microprofile/microprofile-jwt-auth-2.1/microprofile-jwt-auth-spec-2.1.html";
 
-    private static final String METRICS_URI = "https://download.eclipse.org/microprofile/microprofile-metrics-3.0/microprofile-metrics-spec-3.0.html";
-
-    private static final String OPEN_TRACING_URI = "https://download.eclipse.org/microprofile/microprofile-opentracing-2.0/microprofile-opentracing-spec-2.0.html";
-
-    public String getFaultToleranceUri() {
-        return FAULT_TOLERANCE_URI;
-    }
-
-    public String getJwtUri() {
-        return JWT_URI;
-    }
+    private static final String REST_CLIENT_URI = "https://download.eclipse.org/microprofile/microprofile-rest-client-4.0/microprofile-rest-client-spec-4.0.html";
 
     public String getConfigUri() {
         return CONFIG_URI;
+    }
+
+    public String getFaultToleranceUri() {
+        return FAULT_TOLERANCE_URI;
     }
 
     public String getHealthUri() {
         return HEALTH_URI;
     }
 
+    public String getTelemetryUri() {
+        return TELEMETRY_URI;
+    }
+
     public String getOpenApiUri() {
         return OPEN_API_URI;
     }
 
+    public String getJwtAuthenticationUri() {
+        return JWT_AUTHENTICATION_URI;
+    }
+
     public String getRestClientUri() {
         return REST_CLIENT_URI;
-    }
-
-    public String getMetricsUri() {
-        return METRICS_URI;
-    }
-
-    public String getOpenTracingUri() {
-        return OPEN_TRACING_URI;
     }
 }
