@@ -1,7 +1,7 @@
 package com.gepardec.training.microprofile;
 
-import javax.ws.rs.ApplicationPath;
-import javax.ws.rs.core.Application;
+import jakarta.ws.rs.ApplicationPath;
+import jakarta.ws.rs.core.Application;
 
 @ApplicationPath("/api")
 public class MvcApplication extends Application {
