@@ -1,16 +1,13 @@
 package com.gepardec.training.microprofile.advanced;
 
-import com.gepardec.training.microprofile.TrainingMetric;
-
-import javax.enterprise.context.RequestScoped;
-import javax.mvc.Controller;
-import javax.ws.rs.GET;
-import javax.ws.rs.Path;
+import jakarta.enterprise.context.RequestScoped;
+import jakarta.mvc.Controller;
+import jakarta.ws.rs.GET;
+import jakarta.ws.rs.Path;
 
 @Path("/advanced")
 @RequestScoped
 @Controller
-@TrainingMetric
 public class IndexController {
 
     @Path("/")
