@@ -1,20 +1,13 @@
 package com.gepardec.training.microprofile.advanced.faulttolerance;
 
-import org.eclipse.microprofile.metrics.annotation.*;
-
-import javax.enterprise.context.RequestScoped;
-import javax.mvc.Controller;
-import javax.ws.rs.GET;
-import javax.ws.rs.Path;
+import jakarta.enterprise.context.RequestScoped;
+import jakarta.mvc.Controller;
+import jakarta.ws.rs.GET;
+import jakarta.ws.rs.Path;
 
 @Path("/advanced/faulttolerance")
 @RequestScoped
 @Controller
-@Counted(name = "counted")
-@Metered(name = "metered")
-@Timed(name = "timed")
-@SimplyTimed(name = "simply-timed")
-@ConcurrentGauge(name = "concurrent-gauge")
 public class IndexController {
 
     @GET
