@@ -2,12 +2,12 @@ package com.gepardec.training.microprofile.basic.config;
 
 import com.gepardec.training.microprofile.SystemHelper;
 
-import javax.enterprise.context.RequestScoped;
-import javax.inject.Inject;
-import javax.mvc.Controller;
-import javax.mvc.Models;
-import javax.ws.rs.GET;
-import javax.ws.rs.Path;
+import jakarta.enterprise.context.RequestScoped;
+import jakarta.inject.Inject;
+import jakarta.mvc.Controller;
+import jakarta.mvc.Models;
+import jakarta.ws.rs.GET;
+import jakarta.ws.rs.Path;
 
 @Path("/basic/config/env")
 @RequestScoped
