@@ -2,12 +2,12 @@ package com.gepardec.training.microprofile.basic.config;
 
 import com.gepardec.training.microprofile.SystemHelper;
 
-import javax.enterprise.context.RequestScoped;
-import javax.inject.Inject;
-import javax.mvc.Controller;
-import javax.mvc.Models;
-import javax.ws.rs.GET;
-import javax.ws.rs.Path;
+import jakarta.enterprise.context.RequestScoped;
+import jakarta.inject.Inject;
+import jakarta.mvc.Controller;
+import jakarta.mvc.Models;
+import jakarta.ws.rs.GET;
+import jakarta.ws.rs.Path;
 
 @Path("/basic/config/file")
 @RequestScoped
@@ -27,11 +27,11 @@ public class FileController {
     public String getConfigFromPropertyFile() {
         model.put("applicationName", appName);
         if (systemHelper.isEnvVariable("FILE_APPLICATION_NAME")) {
-            model.put("fileMessage", "Warning: Config is provided via a env variable and not via microprofile-properties.properties.");
+            model.put("fileMessage", "Warning: Config is provided via a env variable and not via microprofile-config.properties.");
         } else if (systemHelper.isSystemProperty("file.application.name")) {
-            model.put("fileMessage", "Warning: Config is provided via system property and not via microprofile-properties.properties.");
+            model.put("fileMessage", "Warning: Config is provided via system property and not via microprofile-config.properties.");
         } else if (appName == null) {
-            model.put("fileMessage", "Warning: Config is not provided via microprofile-properties.properties.");
+            model.put("fileMessage", "Warning: Config is not provided via microprofile-config.properties.");
         }
         return "basic/config/file.xhtml";
     }
