@@ -1,18 +1,13 @@
 package com.gepardec.training.microprofile.basic.config;
 
-import com.gepardec.training.microprofile.TrainingMetric;
-import org.eclipse.microprofile.opentracing.Traced;
-
-import javax.enterprise.context.RequestScoped;
-import javax.mvc.Controller;
-import javax.ws.rs.GET;
-import javax.ws.rs.Path;
+import jakarta.enterprise.context.RequestScoped;
+import jakarta.mvc.Controller;
+import jakarta.ws.rs.GET;
+import jakarta.ws.rs.Path;
 
 @Path("/basic/config")
 @RequestScoped
 @Controller
-@TrainingMetric
-@Traced
 public class IndexController {
 
     @Path("/")
