@@ -3,8 +3,8 @@ package com.gepardec.training.microprofile.basic.faulttolerance;
 import org.apache.commons.lang3.concurrent.ConcurrentUtils;
 import org.slf4j.Logger;
 
-import javax.enterprise.context.Dependent;
-import javax.inject.Inject;
+import jakarta.enterprise.context.Dependent;
+import jakarta.inject.Inject;
 import java.util.concurrent.Future;
 
 @Dependent
