@@ -1,7 +1,7 @@
 package com.gepardec.training.microprofile.basic.restclient;
 
-import javax.ws.rs.*;
-import javax.ws.rs.core.MediaType;
+import jakarta.ws.rs.*;
+import jakarta.ws.rs.core.MediaType;
 
 @Path("/")
 public interface HttpbinClientApi {
