@@ -2,7 +2,7 @@ package com.gepardec.training.microprofile.basic.restclient;
 
 import org.slf4j.Logger;
 
-import javax.inject.Inject;
+import jakarta.inject.Inject;
 
 public class RestClientHeaderHandler {
 
